pyproject.toml
[tool.poetry]
name = "xingbot"
version = "0.1.0"
description = "Minimal XING job collector + GPT evaluator + best-effort applier"
authors = ["you <you@example.com>"]
packages = [{ include = "xingbot", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
playwright = ">=1.40"
loguru = "^0.7.2"
pyyaml = "^6.0.2"
openai = ">=1.0.0"
langdetect = "^1.0.9"
httpx = "^0.28.1"

[tool.poetry.scripts]
xingbot = "xingbot.cli:main"
xingbot-auto = "xingbot.auto_run:main"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q --basetemp=.tmp"
testpaths = ["tests"]
markers = [
    "integration: integration tests that use mocked/stubbed async HTTP and/or Playwright",
    "e2e: manual e2e checks (disabled unless explicitly requested)",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.24.0"
respx = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
